fbtty: dynamic scrollback buf size

+++ Kernel/inc/drivers/fb/fb.h
@@ -11,8 +11,8 @@ typedef struct {
 typedef struct {
     fb_info_t *(*get_info)(void);
     void (*setpixel)(int x, int y, uint32_t pixel);
-    void (*readpixels)(uint64_t *buf, size_t offset, size_t count);
-    void (*memcpy)(uint64_t *buf, size_t offset, size_t count);
+    void (*readpixels)(uint32_t *buf, size_t offset, size_t count);
+    void (*memcpy)(uint32_t *buf, size_t offset, size_t count);
 } fb_driver_t;
 
 int fb_register_driver(fb_driver_t *driver);
